Move array-oriented functions to their own file
