+++ src/scyjava/__init__.py
@@ -17,6 +17,11 @@ from jpype.types import (
     JShort,
 )
 
+from scyjava._arrays import (  # noqa: F401
+    is_arraylike,
+    is_memoryarraylike,
+    is_xarraylike,
+)
 from scyjava._java import (  # noqa: F401
     JavaClasses,
     is_awt_initialized,
@@ -134,57 +139,6 @@ def compare_version(version, java_class_version):
     )
     return version != java_class_version and is_version_at_least(
         java_class_version, version
-    )
-
-
-# -- Type reasoning --
-
-
-def is_arraylike(arr):
-    """
-    Return True iff the object is arraylike: possessing
-    .shape, .dtype, .__array__, and .ndim attributes.
-
-    :param arr: The object to check for arraylike properties
-    :return: True iff the object is arraylike
-    """
-    return (
-        hasattr(arr, "shape")
-        and hasattr(arr, "dtype")
-        and hasattr(arr, "__array__")
-        and hasattr(arr, "ndim")
-    )
-
-
-def is_memoryarraylike(arr):
-    """
-    Return True iff the object is memoryarraylike:
-    an arraylike object whose .data type is memoryview.
-
-    :param arr: The object to check for memoryarraylike properties
-    :return: True iff the object is memoryarraylike
-    """
-    return (
-        is_arraylike(arr)
-        and hasattr(arr, "data")
-        and type(arr.data).__name__ == "memoryview"
-    )
-
-
-def is_xarraylike(xarr):
-    """
-    Return True iff the object is xarraylike:
-    possessing .values, .dims, and .coords attributes,
-    and whose .values are arraylike.
-
-    :param arr: The object to check for xarraylike properties
-    :return: True iff the object is xarraylike
-    """
-    return (
-        hasattr(xarr, "values")
-        and hasattr(xarr, "dims")
-        and hasattr(xarr, "coords")
-        and is_arraylike(xarr.values)
     )
